+++ R/googleAuthR.R
@@ -11,8 +11,11 @@
 #' @section Batching:
 #' 
 #' If you have many API calls, you can save a lot of time by using batching.  
-#'   This takes your many calls and sends them in one POST request to /batch, 
+#'   This takes your many calls and sends them in one POST request to \code{www.googleapis.com/batch}, 
 #'   returning a list of any responses.  See \code{\link{gar_batch}} for details.
+#'   
+#' Another common batch task is to call the same function with one parameter changing each call.  
+#'  This is supported using the \code{\link{gar_batch_walk}} function.
 #' 
 #' @section Shiny functions:
 #'